Moved the generator methods that were generating Pulse_Block_Ensemble and Pulse_Block objects out of the GUI table from the sequence_generator_logic to the pulsed GUI module. This step was taken because these methods simply extract information from the GUI table and put them into the dedicated data objects. Therefore this has nothing to do with the functionality of the logic. In order to ensure that GUI and logic now have access to the objects classes, the class definitions have been moved to a separate file named "pulse_objects.py" that is imported into GUI and logic module.
