+++ logic/sequence_generator_logic.py
@@ -29,397 +29,9 @@ from collections import OrderedDict
 import inspect
 import importlib
 
+from logic.pulse_objects import Pulse_Block_Element, Pulse_Block, Pulse_Block_Ensemble, Pulse_Sequence
 from logic.generic_logic import GenericLogic
 from logic.sampling_functions import SamplingFunctions
-
-
-class Pulse_Block_Element(object):
-    """ Object representing a single atomic element in a pulse block.
-
-    This class can build waiting times, sine waves, etc. The pulse block may
-    contain many Pulse_Block_Element Objects. These objects can be displayed in
-    a GUI as single rows of a Pulse_Block.
-    """
-    def __init__(self, init_length_bins, increment_bins=0, pulse_function=None, digital_high=None,
-                 parameters=[], use_as_tick=False):
-        """ The constructor for a Pulse_Block_Element needs to have:
-
-        @param int init_length_bins: an initial length of a bins, this
-                                 parameters should not be zero but must have a
-                                 finite value.
-        @param int increment_bins: the number which will be incremented during
-                               each repetition of this object
-        @param list pulse_function: list of strings with name of the sampling
-                                    function how to alter the points, the name
-                                    of the function will be one of the sampling
-                                    functions
-        @param list digital_high: list of digital channels, which are for the
-                              length of this Pulse_Block_Element are set either
-                              to True (high) or to False (low). The length of
-                              the marker list depends on the number of (active)
-                              digital channels. For 4 digital channel it may
-                              look like:
-                              [True, False, False, False]
-        @param list parameters: a list of dictionaries. The number of dictionaries
-                           depends on the number of analog channels. The
-                           number of entries within a dictionary depends on the
-                           chosen sampling function. The key words of the
-                           dictionary for the parameters will be those of the
-                           sampling functions.
-        @param bool use_as_tick: bool, indicates, whether the set length should
-                           be used as a tick (i.e. the parameter for the x axis)
-                           for the later plot in the analysis.
-        """
-        # FIXME: Sanity checks need to be implemented here
-        self.init_length_bins   = init_length_bins
-        self.increment_bins     = increment_bins
-        self.pulse_function     = pulse_function
-        self.digital_high       = digital_high
-        self.parameters         = parameters
-        self.use_as_tick        = use_as_tick
-
-        # calculate number of digital and analogue channels
-        if pulse_function is not None:
-            self.analog_channels = len(pulse_function)
-        else:
-            self.analog_channels = 0
-        if digital_high is not None:
-            self.digital_channels = len(digital_high)
-        else:
-            self.digital_channels = 0
-
-
-class Pulse_Block(object):
-    """ Collection of Pulse_Block_Elements which is called a Pulse_Block. """
-
-    def __init__(self, name, element_list):
-        """ The constructor for a Pulse_Block needs to have:
-
-        @param str name: chosen name for the Pulse_Block
-        @param list element_list: which contains the Pulse_Block_Element
-                             Objects forming a Pulse_Block, e.g.
-                             [Pulse_Block_Element, Pulse_Block_Element, ...]
-        """
-        self.name = name
-        self.element_list = element_list
-        self.refresh_parameters()
-
-    def refresh_parameters(self):
-        """ Initialize the parameters which describe this Pulse_Block object.
-
-        The information is gained from all the Pulse_Block_Element objects,
-        which are attached in the element_list.
-        """
-        # the Pulse_Block parameter
-        self.init_length_bins = 0
-        self.increment_bins = 0
-        self.analog_channels = 0
-        self.digital_channels = 0
-        self.use_as_tick = False
-
-        # calculate the tick value for the whole block. Basically sum all the
-        # init_length_bins which have the use_as_tick attribute set to True.
-        self.measurement_tick_start = 0
-        # make the same thing for the increment, to obtain the total increment
-        # number for the block. This facilitates in calculating the measurement tick list.
-        self.measurement_tick_increment = 0
-
-        for elem in self.element_list:
-            self.init_length_bins += elem.init_length_bins
-            self.increment_bins += elem.increment_bins
-
-            if elem.use_as_tick:
-                self.use_as_tick = True
-                self.measurement_tick_start += elem.init_length_bins
-                self.measurement_tick_increment += elem.increment_bins
-
-            if elem.analog_channels > self.analog_channels:
-                self.analog_channels = elem.analog_channels
-            if elem.digital_channels > self.digital_channels:
-                self.digital_channels = elem.digital_channels
-
-    def replace_element(self, position, element):
-        self.element_list[position] = element
-        self.refresh_parameters()
-        return
-
-    def delete_element(self, position):
-        del(self.element_list[position])
-        self.refresh_parameters()
-        return
-
-    def append_element(self, element, at_beginning=False):
-        if at_beginning:
-            self.element_list.insert(0, element)
-        else:
-            self.element_list.append(element)
-        self.refresh_parameters()
-        return
-
-
-class Pulse_Block_Ensemble(object):
-    """ Represents a collection of Pulse_Block objects which is called a
-        Pulse_Block_Ensemble.
-
-    This object is used as a construction plan to create one sampled file.
-    """
-
-    def __init__(self, name, block_list, activation_config, laser_channel=None, rotating_frame=True):
-        """ The constructor for a Pulse_Block_Ensemble needs to have:
-
-        @param str name: chosen name for the Pulse_Block_Ensemble
-        @param list block_list: contains the Pulse_Block Objects with their number of repetitions,
-                                e.g.
-                                    [(Pulse_Block, repetitions), (Pulse_Block, repetitions), ...])
-        @param list activation_config: A list of strings representing the channel configuration
-                                        e.g. ['a_ch1', 'd_ch1', 'd_ch2', 'a_ch2']
-        @param str laser_channel: The string descriptor for the laser channel.
-                                    Must be contained by the passed activation_config.
-        @param bool rotating_frame: indicates whether the phase should be preserved for all the
-                                    functions.
-        """
-        # FIXME: Sanity checking needed here
-        self.name = name                    # Pulse_Block_Ensemble name
-        self.block_list = block_list
-        self.activation_config = activation_config
-        self.rotating_frame = rotating_frame
-        if laser_channel in activation_config:
-            self.laser_channel = laser_channel
-        else:
-            self.laser_channel = None
-        self.refresh_parameters()
-
-    def refresh_parameters(self):
-        self.length_bins = 0
-        self.analog_channels = 0
-        self.digital_channels = 0
-        self.number_of_lasers = 0
-        # calculate the tick values for the whole block_ensemble.
-        self.measurement_ticks_list = np.array([])
-
-        for block, reps in self.block_list:
-            # Get number of channels from the block information
-            if block.analog_channels > self.analog_channels:
-                self.analog_channels = block.analog_channels
-            if block.digital_channels > self.digital_channels:
-                self.digital_channels = block.digital_channels
-
-            # Get and set information about the length of the ensemble
-            self.length_bins += (block.init_length_bins * (reps+1) + block.increment_bins * (reps*(reps+1)/2))
-
-            # Calculate the measurement ticks list for this ensemble
-            if block.use_as_tick:
-                start = block.measurement_tick_start
-                incr = block.measurement_tick_increment
-                if incr == 0:
-                    arr = np.array([])
-                else:
-                    arr = np.arange(start, start+(reps+1)*incr, incr)
-                self.measurement_ticks_list = np.append(self.measurement_ticks_list, arr)
-
-            # Calculate the number of laser pulses for this ensemble
-            if self.laser_channel is None:
-                self.number_of_lasers = 0
-            elif 'd_ch' in self.laser_channel:
-                # determine the laser channel index for the corresponding channel
-                digital_chnl_list = [chnl for chnl in self.activation_config if 'd_ch' in chnl]
-                laser_index = digital_chnl_list.index(self.laser_channel)
-                # Iterate through the elements and count laser on state changes (no double counting)
-                laser_on = False
-                for elem in block.element_list:
-                    if elem.digital_high[laser_index] and not laser_on:
-                        self.number_of_lasers += 1
-                        laser_on = True
-                    elif not elem.digital_high[laser_index]:
-                        laser_on = False
-            elif 'a_ch' in self.laser_channel:
-                # determine the laser channel index for the corresponding channel
-                analog_chnl_list = [chnl for chnl in self.activation_config if 'a_ch' in chnl]
-                laser_index = analog_chnl_list.index(self.laser_channel)
-                # Iterate through the elements and count laser on state changes (no double counting)
-                laser_on = False
-                for elem in block.element_list:
-                    if elem.pulse_function[laser_index] == 'DC' and not laser_on:
-                        self.number_of_lasers += 1
-                        laser_on = True
-                    elif elem.pulse_function[laser_index] != 'DC':
-                        laser_on = False
-        return
-
-    def replace_block(self, position, block):
-        self.block_list[position] = block
-        self.refresh_parameters()
-        return
-
-    def delete_block(self, position):
-        del(self.block_list[position])
-        self.refresh_parameters()
-        return
-
-    def append_block(self, block, at_beginning=False):
-        if at_beginning:
-            self.block_list.insert(0, block)
-        else:
-            self.block_list.append(block)
-        self.refresh_parameters()
-        return
-
-# one crucial thing about a list is that you needed to know the correspondance
-# i.e. in which entry which information is situated. Dictionaries assign to a
-# keyword a value, which is much more descriptive and which enables us to carry
-# a bigger information content.
-
-
-class Pulse_Sequence(object):
-    """ Higher order object for sequence capability.
-
-    Represents a playback procedure for a number of Pulse_Block_Ensembles.
-    Unused for pulse generator hardware without sequencing functionality.
-    """
-
-
-    def __init__(self, name, ensemble_param_list, rotating_frame=True):
-        """ The constructor for a Pulse_Sequence objects needs to have:
-
-        @param str name: the actual name of the sequence
-        @param list ensemble_param_list: list containing a tuple of two entries:
-                    (Pulse_Block_Ensemble, seq_param), (Pulse_Block_Ensemble, seq_param), ...
-                                          The seq_param is a dictionary, where the various sequence
-                                          parameters are saved with their keywords and the
-                                          according parameter (as item). What parameter will be in
-                                          this dictionary will completely depend on the sequence
-                                          parameter set of the pulsing device. But most certain the
-                                          parameter 'reps' meaning repetitions will be presesnt in
-                                          the sequence parameters.
-                                          If only 'reps' are in the dictionary, than the dict will
-                                          look like
-                                                seq_param = {'reps': 12}
-                                          if 12 was chosen as the number of repetitions.
-        @param bool rotating_frame: indicates, whether the phase has to be preserved in all
-                                    oscillating functions.
-        """
-
-        self.name = name
-        self.ensemble_param_list = ensemble_param_list
-        self.rotating_frame = rotating_frame
-        self.refresh_parameters()
-        self.sampled_ensembles = OrderedDict()
-
-        #TODO: That parameter container should be used to store all Information
-        #      upon creation of the Pulse_Sequence object. Those information
-        #      will also be saved to file.
-        self.param_container = OrderedDict()
-
-    def refresh_parameters(self):
-        """ Generate the needed parameters from the passed object.
-
-        Baiscally, calculate the length_bins and number of analog and digital
-        channels.
-        """
-        self.length_bins = 0
-        self.analog_channels = 0
-        self.digital_channels = 0
-        # here all DIFFERENT kind of ensembles will be saved in, i.e. with different names.
-        # REMEMBER: A dict is hashable, that means, you will get its entry with the complexity of 1
-        #           without searching through the whole dict!
-        self.different_ensembles_dict = dict()
-
-        # here the measurement ticks will be saved:
-        self.measurement_ticks_list = []
-
-        self.number_of_lasers = 0
-        # to make a resonable measurement tick list, the last biggest tick value after all
-        # the repetitions of a block is used as the offset_time for the next
-        # block.
-        offset_tick_bin = 0
-
-        for ensemble, seq_dict in self.ensemble_param_list:
-
-            for param in seq_dict:
-                if 'reps' in param.lower() or 'repetition' in param.lower():
-                    reps = seq_dict[param]
-                    break
-                else:
-                    reps = 0
-
-            self.length_bins += (ensemble.length_bins * (reps+1))
-
-            if ensemble.analog_channels > self.analog_channels:
-                self.analog_channels = ensemble.analog_channels
-            if ensemble.digital_channels > self.digital_channels:
-                self.digital_channels = ensemble.digital_channels
-
-            if self.different_ensembles_dict.get(ensemble.name) is None:
-                self.different_ensembles_dict[ensemble.name] = ensemble
-
-            self.measurement_ticks_list = np.append(self.measurement_ticks_list,
-                                                   (offset_tick_bin + ensemble.measurement_ticks_list))
-
-            self.number_of_lasers += ensemble.number_of_lasers
-
-            # for the next repetition or pulse_block_ensemble, add last number
-            # from the measurement_ticks_list as offset_tick_bin. Otherwise the
-            # measurement_ticks_list will be a mess:
-            if len(self.measurement_ticks_list) > 0:
-                offset_tick_bin = self.measurement_ticks_list[-1]
-
-        self.estimated_bytes = self.length_bins * (self.analog_channels * 4 + self.digital_channels)
-
-    def replace_ensemble(self, position, ensemble_param):
-        """ Replace an ensemble at a given position.
-
-        @param int position: position in a the ensemble list
-        @param list ensemble_param: with entries
-                                        (Pulse_Block_Ensemble, seq_param)
-                                    which will replace the old one.
-        """
-        self.ensemble_param_list[position] = ensemble_param
-        self.refresh_parameters()
-        return
-
-    def delete_ensemble(self, position):
-        """ Delete an ensemble at a given position
-
-        @param int position: position within the list self.ensemble_param_list.
-        """
-        del(self.ensemble_list[position])
-        self.refresh_parameters()
-
-    def append_ensemble(self, ensemble_param, at_beginning=False):
-        """ Append either at the front or at the back an ensemble_param
-
-        @param tuple ensemble_param: containing two entries:
-                                        (Pulse_Block_Ensemble, seq_param)
-                                     where Pulse_Block_Ensemble is the object
-                                     and seq_param is the parameter set for that
-                                     ensemble.
-        @param bool at_beginning: If flase append to end (default), if true then
-                                  inset at beginning.
-        """
-
-        if at_beginning:
-            self.ensemble_list.insert(0, ensemble_param)
-        else:
-            self.ensemble_list.append(ensemble_param)
-        self.refresh_parameters()
-
-    #TODO: Experimental, check how necessary that method is and replace by other idea if needed:
-    def set_sampled_ensembles(self, sampled_dict):
-        """ Set the dict, containing the sampled ensembles of the sequence.
-
-        @param dict sampled_dict: the ordered dictionary with a list of strings telling about the
-                                  actual names of the created sequences
-        """
-        self.sampled_ensembles = sampled_dict
-
-    # TODO: Experimental, check how necessary that method is and replace by other idea if needed:
-    def get_sampled_ensembles(self):
-        """ Retrieve the dict, which tells into which names the ensembles have been sampled
-
-        @return dict: a dict with keys being the ensemble names and items being a list of string
-                      names, which were created by the hardware.
-        """
-        return self.sampled_ensembles
 
 class SequenceGeneratorLogic(GenericLogic, SamplingFunctions):
     """unstable: Nikolas Tomek
@@ -627,29 +239,29 @@ class SequenceGeneratorLogic(GenericLogic, SamplingFunctions):
 
     # ========================================================================
 
-    def get_func_config(self):
-        """ Retrieve func_config dict of the logic, including hardware constraints.
-
-        @return dict: with all the defined functions and their corresponding
-                      parameters and constraints.
-
-        The contraints from the hardware are now also included in the dict. How
-        the returned dict is looking like is defined in the inherited class
-        SamplingFunctions.
-        """
-        func_config = self.func_config
-
-        # set the max amplitude from the hardware:
-        # ampl_max = const['amplitude_analog'][1]
-        # FIXME: You see it... below should be the actual amplitude constraint instead of 0.5
-        ampl_max = 0.5/2.0
-        if ampl_max is not None:
-            for func in func_config:
-                for param in func_config[func]:
-                    if 'amplitude' in param:
-                        func_config[func][param]['max'] = ampl_max
-
-        return func_config
+    # def get_func_config(self):
+    #     """ Retrieve func_config dict of the logic, including hardware constraints.
+    #
+    #     @return dict: with all the defined functions and their corresponding
+    #                   parameters and constraints.
+    #
+    #     The contraints from the hardware are now also included in the dict. How
+    #     the returned dict is looking like is defined in the inherited class
+    #     SamplingFunctions.
+    #     """
+    #     func_config = self.func_config
+    #
+    #     # set the max amplitude from the hardware:
+    #     # ampl_max = const['amplitude_analog'][1]
+    #     # FIXME: You see it... below should be the actual amplitude constraint instead of 0.5
+    #     ampl_max = 0.5/2.0
+    #     if ampl_max is not None:
+    #         for func in func_config:
+    #             for param in func_config[func]:
+    #                 if 'amplitude' in param:
+    #                     func_config[func][param]['max'] = ampl_max
+    #
+    #     return func_config
 
     def set_activation_config(self, activation_config):
         """
@@ -949,119 +561,6 @@ class SequenceGeneratorLogic(GenericLogic, SamplingFunctions):
         self.saved_pulse_sequences = sequences
         self.signal_sequence_list_updated.emit()
         return
-
-    def generate_pulse_block_object(self, pb_name, block_matrix):
-        """ Generates from an given table block_matrix a block_object.
-
-        @param pb_name: string, Name of the created Pulse_Block Object
-        @param block_matrix: structured np.array, matrix, in which the
-                             construction plan for Pulse_Block_Element objects
-                             are displayed as rows.
-
-        Three internal dict where used, to get all the needed information about
-        how parameters, functions are defined (_add_pbe_param,func_config and
-        _unit_prefix).
-        The dict cfg_param_pbe (configuration parameter declaration dict for
-        Pulse_Block_Element) stores how the objects are appearing in the GUI.
-        This dict enables the proper access to the desired element in the GUI.
-        """
-
-        # list of all the pulse block element objects
-        pbe_obj_list = [None]*len(block_matrix)
-
-        # seperate digital and analogue channels
-        analog_chnl_names = [a_chnl for a_chnl in self.activation_config if 'a_ch' in a_chnl]
-        digital_chnl_names = [d_chnl for d_chnl in self.activation_config if 'd_ch' in d_chnl]
-
-        for row_index, row in enumerate(block_matrix):
-            # check how length is displayed and convert it to bins:
-            length_time= row[self.cfg_param_pbe['length']]
-            init_length_bins=int(np.round(length_time*self.sample_rate))
-
-            # check how increment is displayed and convert it to bins:
-            increment_time=row[self.cfg_param_pbe['increment']]
-            increment_bins= int(np.round(increment_time*self.sample_rate))
-
-            # get the dict with all possible functions and their parameters:
-            func_dict = self.get_func_config()
-
-            # get the proper pulse_functions and its parameters:
-            pulse_function=[None]*self.analog_channels
-            parameter_list =[None]*self.analog_channels
-            for num in range(self.analog_channels):
-                # get the number of the analogue channel according to the channel activation_config
-                a_chnl_number = analog_chnl_names[num].split('ch')[-1]
-                pulse_function[num] = row[self.cfg_param_pbe['function_'+a_chnl_number]].decode('UTF-8')
-
-                # search for this function in the dictionary and get all the
-                # parameter with their names in list:
-                param_dict = func_dict[pulse_function[num]]
-                parameters = {}
-                for entry in list(param_dict):
-                    # Obtain how the value is displayed in the table:
-                    param_value = row[self.cfg_param_pbe[entry+'_'+a_chnl_number]]
-                    parameters[entry] = param_value
-                parameter_list[num] = parameters
-
-            digital_high = [None]*self.digital_channels
-            for num in range(self.digital_channels):
-                # get the number of the digital channel according to the channel activation_config
-                d_chnl_number = digital_chnl_names[num].split('ch')[-1]
-                digital_high[num] = bool(row[self.cfg_param_pbe['digital_'+d_chnl_number]])
-
-            use_as_tick = bool(row[self.cfg_param_pbe['use']])
-
-            # create here actually the object with all the obtained information:
-            pbe_obj_list[row_index] = Pulse_Block_Element(
-                        init_length_bins=init_length_bins,
-                        increment_bins=increment_bins,
-                        pulse_function=pulse_function,
-                        digital_high=digital_high,
-                        parameters=parameter_list,
-                        use_as_tick=use_as_tick)
-
-        pb_obj = Pulse_Block(pb_name, pbe_obj_list)
-        self.save_block(pb_name, pb_obj)
-        self.current_block = pb_obj
-
-
-    def generate_pulse_block_ensemble(self, ensemble_name, ensemble_matrix, rotating_frame=True):
-        """
-        Generates from an given table ensemble_matrix a ensemble object.
-
-        @param str ensemble_name: Name of the created Pulse_Block_Ensemble object
-        @param np.array ensemble_matrix: structured 2D np.array, matrix, in which the construction
-                                         plan for Pulse_Block objects are displayed as rows.
-        @param str laser_channel: the channel controlling the laser
-        @param bool rotating_frame: optional, whether the phase preservation is mentained
-                                    throughout the sequence.
-
-        The dict cfg_param_pb (configuration parameter declaration dict for Pulse_Block) stores how
-        the objects are related to each other in a sequencial way. That relationship is used in the
-        GUI, where the parameters appear in columns.
-        This dict enables the proper access to the desired element in the GUI.
-        """
-
-        # list of all the pulse block element objects
-        pb_obj_list = [None]*len(ensemble_matrix)
-
-        for row_index, row in enumerate(ensemble_matrix):
-            pulse_block_name = row[self.cfg_param_pb['pulse_block']].decode('UTF-8')
-            pulse_block_reps = row[self.cfg_param_pb['repetition']]
-            # Fetch previously saved block object
-            block = self.get_pulse_block(pulse_block_name)
-            # Append block object along with repetitions to the block list
-            pb_obj_list[row_index] = (block, pulse_block_reps)
-
-        # Create the Pulse_Block_Ensemble object
-        pulse_block_ensemble = Pulse_Block_Ensemble(name=ensemble_name, block_list=pb_obj_list,
-                                                    activation_config=self.activation_config,
-                                                    laser_channel=self.laser_channel,
-                                                    rotating_frame=rotating_frame)
-        # set current block ensemble
-        self.current_ensemble = pulse_block_ensemble
-        # save ensemble
-        self.save_ensemble(ensemble_name, pulse_block_ensemble)
 
     #---------------------------------------------------------------------------
     #                    END sequence/block generation
